Model: TC: Update model & RSS firmware in step
As it happens, there was inverted logic in TF-M's rss
that would check that a particular KMU interrupt has not fired.
After correcting this logic, I found that the firmware no longer
booted, perpetually waiting for the KMU interrupt that the model
did not provide.

This update resolves the issue with the model.

New RSS Firmware has a slightly different flash map, so this includes
the modifications needed to correct image offsets.

Further, Update RSS firmware allowing rollback

In particular, this introduces a subdirectory within the TC prebuilt dir
that corresponds to the short hash of the TF-M revision used to build
the rss images. This allows a commit to the CI in the future to change
both the model version and the rss version in step, without replacing
the old revision. This should allow the CI to test any single commit
without worrying about changing the model version without also changing
the rss version and vice versa.

Change-Id: I0c0f0e7036741f0a9a8cabed605ed21984a691fb
Signed-off-by: Jimmy Brisson <jimmy.brisson@arm.com>