create-model-dockerfile: update ubuntu to jammy
Currently tests are run on Ubuntu Bionic. Released in 2018, it is
starting to show its age and packages are becoming very out of date.
This is mostly apparent with python, where packages are dropping support
for versions under 3.7 while only 3.6 is available on Bionic.

Update the used image to Jammy. There already exists a dockerfile for it
so this change just enables it.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I5a76572ab714a9dc355fbd2f741747b8a212c743