tf-cov-make: avoid usage of parenthesis with local variable
For bash'ism $(..) is command substitution method so output of
command is substituted, here to determine output of the command
subshell maybe used and thus can lead to not able to resolve local
variables if they are used inside parenthesis.

Use braces {..} as its a parameter expansion and doesn't involve
subshell.

The issue is observed in this job[1]:
[..]
    	tf-cov-make: ....
	stm32mp1_common_flags: command not found
[..]

[1]:
https://ci.trustedfirmware.org/job/tf-a-coverity/499/console

Change-Id: Ida7a39f25623213fa82374eeb5d9fd02d5d1192a
Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>