ci(static-checks): improve checks for thirdparty headers
current include order check doesn't consider how third-party headers
should be included.

With - https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/19825

We are proposing a approach for TF-A to use this include order
for any third-party headers.

Order proposed is:

[..]
 /* system header files */
  #include <stddef.h>

  /* Thirdparty headers - mbedtls header files */
  #include <mbedtls/version.h>

  /* project header files */
  #include <drivers/auth/auth_mod.h>
  #include <drivers/auth/tbbr_cot_common.h>

  /* platform header files */
  #include <platform_def.h>
[..]

Use this same order for CI static checks for include order tests.

Change-Id: I4637e1ed839dafbbfa8039972ccca79059ca8897
Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>