tf-coverity: exclude Juno files for host tools
When the Arm(R) Ethos(TM)-N NPU driver is built with TZMP1 support
enabled for the Juno platform, there are platform specific files used
for building the host tools.

Files for building the host tools are excluded from Coverity scans but
as these are Juno specific files they are not covered by those
exclusions.

To resolve this, new exclusions have been added for the Juno specific
files for building the host tools.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>
Change-Id: I9a3657c8f1dd0133d309d97b123690ba71e747d1