chore: remove no-<feat> negative tests
TF-A has a policy of if a feature enable flag is set to 0 then all
feature related code should be skipped. TF-A also has a policy that all
features present in hardware must also be enabled. So configuration
where the feature flag is set to 0 but the test includes it in hardware
is invalid. Additionally, since enablement is skipped, no particular
behaviour can be expected.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I9728d69f7bf95d7ffb4df41fc3c722b0602a47ac