ci(n1sdp): refactor fip fragment for specificity
The ``n1sdp-fip`` fragment builds the fip image with UEFI. Replace the
fragment to target this specific use case as we also want to generate
fip images with other NS workloads i.e. TFTF. Move parts of the code for
creating a fip image with xCP firmware to a generic location so it may
be used by other fragments.

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: I37433e32f1bc235e4c524ceed5730b0bc4c63ddd