chore: make a feature all-config
Each feature is currently tested in its own config, requiring a full
build and a run of a bunch of unrelated tests which is inefficient and
slows down CI. Since TF-A's fvp build (fvp-default) now enables all
features dynamically by default, this can happen in 1 single config.
Introduce it and remove all others.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: Iaa227664e3da26594e45e4b763f1d5dfca0f9e9c