utils.sh: fetch_file: Increase timeout/retries
As a reaction to https://ci.trustedfirmware.org/job/tf-a-builder/2072683/console :

00:00:43.364 + curl --fail --connect-timeout 5 --retry 5 -sLS -o /home/buildslave/workspace/tf-a-builder/workspace/mbedtls.tar.gz https://github.com/Mbed-TLS/mbedtls/archive/mbedtls-2.28.1.tar.gz
00:00:48.381 curl: (28) SSL connection timeout
00:05:20.165 curl: (56) OpenSSL SSL_read: Connection reset by peer, errno 104

So, based on the log, it doesn't even seem that curl made any retries. Reported
as https://github.com/curl/curl/issues/11124 .

In the meantime, let's hope that these increases help anyhow, if not, we'll
need to add retrying on our side.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I23481d4fcac155964343cef588e04474eec7678c