gen_test_desc.py: Allow to temporarily deactivate some configs in a group
For this, a config should be renamed to have ".inactive" suffix. This is
apparently better than removing it altogether. This feature is needed in
case there's an issue with build/test process of a particular config
which jeopardize/break the entire CI jobs (in particular, what specifically
prompted addition of this feature is a particular device being offline in
LAVA, so all test jobs submitted just hanged in queue, while Jenkins builds
waiting for them until timeout hit, so entire top-level jobs were guaranteedly
failing, hogging builder agents for hours).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I164a099350f452aa3349df5f2442b1b2bbbd8d8e