ci(static-checks): skip ignored files when checking patches include order
The script that checks include order skips some files and directories,
when run in a folder. This should be the same when using --patch option.
Use the same utils.file_is_ignored() in inc_order_is_correct() to skip
files.
This was was triggered when checking some device tree files, for which
the order of include is important, but do not match TF-A rules for
include order.

Signed-off-by: Yann Gautier <yann.gautier@st.com>
Change-Id: Idda3c2ad4b10803128396fcf3695ae6139850d2d