fix: Quiesce confusing error message from apply_patch()
Some tests require deploying a local patch on TF-A sources prior to
compiling it. This works in 2 steps:
1) Check if the patch is already applied.
2) If no, apply it.

In the (most likely) event where the patch needs applying, a confusing
error message is printed right now:

  Applying ubsan/ubsan_test.patch...
  ++ git apply --reverse --check
  error: patch failed: bl31/bl31_main.c:108
  error: bl31/bl31_main.c: patch does not apply

This error message is issued by the git command checking whether the
patch can be *reverted*, in other words that the patch is *not*
applied yet, which is the normal expectation.

Silence this error message to avoid any confusion. In the (unlikely)
event where the patch has already been applied, proper messages
continue to be emitted:

  Applying ubsan/ubsan_test.patch...
  Skipping already applied ubsan/ubsan_test.patch

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>
Change-Id: If3089cdb8593af310aa7b99c2cfe3e8647d72cf9