feat(test-configs): add a test config to exercise TF-A's SAVE_KEYS feature
Add a new run config which generates boot certificates and keys, and
saves the latter on the disk.

This is the first test config to leverage TF-A's SAVE_KEYS feature,
which requests the certificate generation tool to save private keys in
PEM format in files, whose names are specified through *_KEY build
options (e.g. BL31_KEY).

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>
Change-Id: I41ed741ef107d8ea638276678521e953b4726b78