chore(coverity): update help message
When asked to download the Coverity Build tool (through the
'--get-coverity-tool' flag), the 'run_coverity_on_tf.sh' script prints a
help message after it has successfully downloaded and unpacked the
tool's tarball.

Update this help message to reflect the latest version of the
tool (2023.6.2) so users should be able to directly copy and paste the
sample command in their shell.

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>
Change-Id: I7009439ed04a218d6eaaaf480dc089a44aef1386