fix(coverity): workaround to fix issue with non utf-8 characters
After upgrading Coverity scan to version 2023.06, we have been facing
several build errors as follows:

[CAPTURE] ERROR: Cannot stat posix_spawn file /tmp/cov-buildslave/1919f32<... non ascii chars... >
[ERROR] Unable to setup argv for posix_spawn cov-translate invocation.
cov-translate died by signal: 6

The workaround suggested by Synopsys support team was to remove the
code that parses the log file leading to the above fail. This does
not impact the analysis of the tarball uploaded to coverity portal
for offline static check analysis.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I5b9ecb12b4165c50ab5fdf72682804305a2fb4f9